feat: added LogGroup monitoring construct
