+++ src/index.ts
@@ -1,3 +1,4 @@
 export * from './PermissionsBoundaryAspect';
 export * from './DeadLetterQueue';
+export * from './ErrorMonitoringAlarm';
 export * from './LambdaMonitoringAlarm';
